Add a max-retry time to avoid curl buildup in healthcheck
We have observed that our healthcheck commands were **not** being killed
when the route_registrar's timeout expired, which allowed curl processes
to build up on the api vm.
(See ca8c4c7cbd15bd1c642865cc8b02beaa91873c74)

To avoid this, this commit adds a `--max-retry-timeout` flag that will
stop retrying when it expires. This commit allows operators to configure
this value, which should be in sync with the route_registrar's api
healthcheck timeout value.

[#158990935]

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>