No longer recursively-`chown` the blobstore files directory
We started doing this about 2 years ago to aid the migration from
running the blobstore as `root` to running it as `vcap`.  We don't
anticipate that this is going to be an issue anymore, and we have had
several reports from production deployers that the recursive `chown`
takes an inordinate amount of time.

In the event that there is a file that does not have the correct owner,
the error message is quite clear - see #107 for an example.

[Finishes #162391689]