Make log timestamp format of tps_watcher and cc_uploader configurable
A previous commit made the tps_watcher and cc_uploader begin logging
timestamps in the RFC3339 format. This change was made in the Diego
components last year and made the default configuration in cf-deployment
in the following commit:
https://github.com/cloudfoundry/cf-deployment/commit/d06a37c1b327104044fceec1fd19b0d502e0fa0b

Since those other platform components have already made the switch, we
think it is unsurprising for the the tps-watcher and cc-uploader
component logs to align with them.

However, to accomodate operators that may need to update their external log aggregators,
this change allows the log format to be switched back to unix-epoch

For additional context:
Diego's cf-dev proposal: https://lists.cloudfoundry.org/g/cf-dev/message/7822
capi-release Pull Request #117

[finishes #162937883]