Call restart_drain when ccng_monit_http_healthcheck fails
The current dependency chain in monit is:

  nginx_cc -> cloud_controller_ng -> ccng_monit_http_healthcheck

The health check is defined as base process so that a failure triggers a restart of the dependent processes.

This is now changed to the logical dependency chain, i.e. the order in which the processes should be started:

  ccng_monit_http_healthcheck -> nginx_cc -> cloud_controller_ng

This removes the implicit restart trigger for the other processes in case the health check fails; but we can add a directive to explicitly invoke the 'restart_drain' script instead.