Make route registration health check fail on 4xx/5xx
The Cloud Controller can return 4xx/5xx responses for a number of
reasons, some of which effectively mean the CC is unavailable (e.g.
nginx is returning 502 because the CC socket is not found). Adding the
`--fail` flag to cURL means that cURL will exit 22 if the /healthz
endpoint has not returned a 2xx/3xx response after all the retries. This
means that the route will not be registered/the heartbeat will fail and
traffic will not be routed to the failing CC VM.

There is likely some similar work to do in the CC monit health check
script to try and automatically repair the CC process when it is failing
but this check should at least make sure users aren't routed to an
unhealthy VM