Notable changes:
* Nginx has been configured to listen on a second port (controlled by
`cc.prom_metrics_server_tls_port`) and use a special set of certs
(controlled by `cc.prom_scraper_tls`) to communicate with the
`prom_scraper`. `prom_scraper` requires that the certificate of the
server that it's scraping is signed with the same CA cert that every
other endpoint in the deployment that it is scraping is signed with. So
-rather than massively complicate configuration for our existing
server configuration- we have decided to create a second server with its
own configuration.

Co-authored-by: Kenneth Lakin <klakin@vmware.com>
Co-authored-by: Michael Oleske <moleske@pivotal.io>
Co-authored-by: David Alvarado <alvaradoda@vmware.com>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>