Adds health check script for Bosh DNS
This commit resolves the core problem in the CC-112 Jira ticket. It
-effectively- instructs Bosh DNS to run the same Cloud Controller health
checking script that `route-registrar` uses to decide whether CCNG's
route should be removed or reinstated in Gorouter.

Because these two health check scripts are run by two separate
components, each with their own start time, and each with its own
frequency with which it will collect and report health check
information, there will inevitably be gaps between the time CCNG is
reported as unhealthy in one system and it is reported as unhealthy in
the other. This seems to be unavoidable.

Co-authored-by: David Alvarado <alvaradoda@vmware.com>
Co-authored-by: Kenneth Lakin <klakin@vmware.com>