Add mariadb_config wrapper script to fix symlink paths
Upgrading from mariadb connector 2.x to 3.x the mariadb_config
started returning absolute paths rather than relative paths to the
mariadb package folders. These absolute paths change when exporting
the release from one director to another.

There doesn't seem to be a set of flags we can manually pass to build
the mysql2 gem so we were forced to add a wrapper script that replaces
the absolute paths with the relative paths.

Co-authored-by: Seth Boyles <sboyles@pivotal.io>
Co-authored-by: Joseph Palermo <jpalermo@pivotal.io>