Update post-backup-unlock.sh.erb (#289)
* Update post-backup-unlock.sh.erb

If the timeout fails then the workers never get started, but monit will eventually restart the web process if the CF install eventually recovers, leaving a VM that is half working (with an unhealthy bosh state) after the script runs.

We could also change the exit behavior of the time out with `set +x` (or is it e? I forget), but it would seem that the only point of timing out is to alert the operator to a possible issue since the CC API can still restart.

* Update post-backup-unlock.sh.erb

Putting the sleep back in after reviewing this [conversation](https://github.com/cloudfoundry/capi-release/pull/132). It seems to be have been added to allow time for route propagation

---------

Co-authored-by: MerricdeLauney <mdelauney@vmware.com>