Remove nginx from response headers and error responses
We observed that the server name (Nginx) is leaked in the header and in the body of an error message.

To enhance the security posture of the web application and mitigate the risk associated with information
disclosure, it is strongly recommended to not share the server name and/or version in any response
information. This does include any type of error messages.

Solution: adjust NGINX sources by using sed to perform an in-place substitution of the server name before building nginx